requirements.txt
aliyun-python-sdk-core-v3==2.8.9
aliyun-python-sdk-ecs==4.9.0
aliyunsdkcore==1.0.2
amqp==1.4.9
anyjson==0.3.3
APScheduler==2.1.2
async-timeout==3.0.0
attrs==18.1.0
autopep8==1.3.5
Babel==2.5.3
billiard==3.3.0.23
boto3==1.4.8
botocore==1.8.50
cchardet==2.1.1
celery==3.1.25
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-auth-ldap==1.2.16
django-celery==3.2.2
django-cors-headers==2.2.0
django-ldapdb==0.9.0
django-silk==2.0.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
docutils==0.14
flower==0.9.2
gevent==1.3.6
gprof2dot==2016.10.13
greenlet==0.4.14
grequests==0.3.0
gunicorn==19.7.1
idna==2.6
idna-ssl==1.1.0
influxdb==5.2.0
Jinja2==2.10
jmespath==0.9.3
kombu==3.0.37
ksc-sdk-python==1.1.1
ldap3==2.5
MarkupSafe==1.0
multidict==4.3.1
olefile==0.44
passlib==1.7.1
Pillow==4.3.0
pkcs7==0.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycares==2.3.0
pycodestyle==2.4.0
pycrypto==2.6.1
pycryptodome==3.6.1
Pygments==2.2.0
PyJWT==1.6.1
pyldap==2.4.37
PyMySQL==0.7.11
python-dateutil==2.7.2
pytz==2017.3
PyYAML==3.11
raven==6.3.0
redis==2.10.6
requests==2.18.4
requests-aws4auth==0.9
s3transfer==0.1.13
six==1.11.0
sqlparse==0.2.4
tornado==5.0.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.11.0
yarl==1.2.6
psycopg2==2.7.5
